- bash script for starting robin_bridge_generated - added restart_codesys arg to run launch

+++ launch/run.sh
@@ -1,21 +1,28 @@
 #!/bin/bash
 
-echo
-echo "[ INFO]: Stopping codesyscontrol service ... Trying with systemctl"
-sudo systemctl stop codesyscontrol
-if [ $? -eq 0 ]; then
+restartCodesys=$1
+
+echo "[ INFO]: Force Codesyscontrol restart: ${restartCodesys}"
+
+if [ $restartCodesys -eq 1 ]; then
     echo
-    echo "[ INFO]: Starting codesyscontrol service via systemctl"
-    sudo systemctl start codesyscontrol
-else
-    echo
-    echo "[ INFO]: Stopping codesyscontrol service ... Trying with service"
-    sudo service codesyscontrol stop
-    echo
-    echo "[ INFO]: Starting codesyscontrol service via service"
-    sudo service codesyscontrol start
+    echo "[ INFO]: Stopping codesyscontrol service ... Trying with systemctl"
+    sudo systemctl stop codesyscontrol
+    if [ $? -eq 0 ]; then
+        echo
+        echo "[ INFO]: Starting codesyscontrol service via systemctl"
+        sudo systemctl start codesyscontrol
+    else
+        echo
+        echo "[ INFO]: Stopping codesyscontrol service ... Trying with service"
+        sudo service codesyscontrol stop
+        echo
+        echo "[ INFO]: Starting codesyscontrol service via service"
+        sudo service codesyscontrol start
+    fi
+    sleep 10
 fi
-sleep 10
+
 if [ $? -eq 0 ]; then
     echo
     echo "[ INFO]: Starting robin bridge ..."
